refactor system & dataset validation again, for good!
