+++ README.md
@@ -25,7 +25,7 @@ pipenv run cli\antoine CHOL
 pipenv run cli\vle CPF CPOL
 pipenv run cli\gamma CHF CHOL -d 25kPa --plot
 pipenv run cli\rk CHF CHOL -d 25kPa,40kPa --plot
-pipenv run cli\herington CHF CHOL -d 25kPa,40kPa --plot
+pipenv run cli\herington CHF CHOL -d 25kPa,40kPa
 pipenv run cli\fredenslund CHF CHOL -d 25kPa --ge --res
 pipenv run cli\fit CHF CHOL -d 25kPa -m NRTL --plot
 ```
