+++ appPy/src/utils/datasets.py
@@ -1,30 +1,20 @@
+import os
+import numpy as np
 from src.utils.errors import AppException
-from src.utils.get_VLE_data import list_VLE_tables
-from src.utils.echo import warn_echo
+from src.utils.systems import validate_system_or_swap, get_system_path
+from src.utils.tsv import open_tsv
 
 
 # get datasets of a binary system, either in the given order, or swapped around
-def get_all_datasets(compound1, compound2):
-    # first try in the given order
-    try:
-        all_dataset_names = list_VLE_tables(compound1, compound2)
+def get_all_dataset_names(compound1, compound2):
+    system_dir_path = get_system_path(compound1, compound2)
+    return [f.name.replace('.tsv', '') for f in os.scandir(system_dir_path) if f.name.endswith('.tsv')]
 
-    # then try to swap it around with warning
-    except AppException as err1:
-        try:
-            (compound1, compound2) = (compound2, compound1)
-            all_dataset_names = list_VLE_tables(compound1, compound2)
-            warn_echo(f'WARNING: compounds were swapped as {compound1}-{compound2} (that system was found)\n')
-
-        # but if nothing is found either, throw the original error, not the swapped one
-        except AppException as err2:
-            raise AppException(err1) from err2
-
-    return all_dataset_names
 
 # for a binary system, parse the 'datasets' comma-separated string and return a list of valid dataset names
+# assumes valid system
 def parse_datasets(compound1, compound2, datasets):
-    all_dataset_names = get_all_datasets(compound1, compound2)
+    all_dataset_names = get_all_dataset_names(compound1, compound2)
 
     if datasets:
         dataset_names = list(filter(bool, map(lambda str: str.strip(), datasets.split(','))))
@@ -33,15 +23,16 @@ def parse_datasets(compound1, compound2, datasets):
         for dataset_name in dataset_names:
             validate_dataset(compound1, compound2, dataset_name, all_dataset_names)
         return dataset_names
-    else:
-        return all_dataset_names
+    return all_dataset_names
+
 
 # wrapper for parse_datasets that fires callback on each valid dataset name
+# may swap compounds order if needed
 def do_datasets(compound1, compound2, datasets, do_for_dataset):
+    compound1, compound2 = validate_system_or_swap(compound1, compound2)
     dataset_names = parse_datasets(compound1, compound2, datasets)
     for dataset_name in dataset_names:
         do_for_dataset(compound1, compound2, dataset_name)
-
 
 
 # throw if dataset does not exist in the system
@@ -49,3 +40,12 @@ def validate_dataset(compound1, compound2, dataset, all_dataset_names):
     if not dataset in all_dataset_names:
         message = f'the dataset {dataset} was not found in system {compound1}-{compound2}!\nAvailable datasets: {", ".join(all_dataset_names)}'
         raise AppException(message)
+
+
+# get specific dataset as a np matrix with columns p/kPa, T/K, x1, y1
+def get_dataset_VLE_table(compound1, compound2, dataset):
+    system_dir_path = get_system_path(compound1, compound2)
+    filename = dataset + '.tsv'
+
+    table = open_tsv(os.path.join(system_dir_path, filename))
+    return np.array(table[1:], dtype='float64')
