--- appPy/src/utils/get_VLE_data.py
@@ -1,32 +1,0 @@
-import os
-import numpy as np
-from .errors import AppException
-from .tsv import open_tsv
-
-
-# helper to get directory path where data is stored for the given compounds system dir, and raise AppException if not exists
-def get_system_path(compound1, compound2):
-    system_name = f'{compound1}-{compound2}'
-    system_dir_path = os.path.join('data', 'VLE', system_name)
-    if not os.path.exists(system_dir_path):
-        system_dir_names = [f.name for f in os.scandir(os.path.join('data', 'VLE')) if f.is_dir()]
-        raise AppException(
-            f'The system {system_name} does not exist in your data!!\nAvailable systems: {", ".join(system_dir_names)}')
-    return system_dir_path
-
-
-# for the given system code, get collection of available dataset names
-def list_VLE_tables(compound1, compound2):
-    system_dir_path = get_system_path(compound1, compound2)
-    return [f.name.replace('.tsv', '') for f in os.scandir(system_dir_path) if f.name.endswith('.tsv')]
-
-
-# get specific dataset as a np matrix with columns p/kPa, T/K, x1, y1
-def get_VLE_table(compound1, compound2, dataset):
-    system_dir_path = get_system_path(compound1, compound2)
-    filename = dataset + '.tsv'
-
-    table = open_tsv(os.path.join(system_dir_path, filename))
-    table = np.array(table[1:], dtype='float64')
-
-    return table
